fix: cleanup startup log levels
